requirements.txt
Django==5.2.6
torch
torchvision
transformers
Pillow
